backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Mail==0.9.1
numpy>=1.26.0,<2.0
pandas>=2.2.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
requests>=2.31.0
yfinance==0.2.66
